Fix typo
The Init of f16 has no effect; Since it's proof of concept, state is copied as float8 every time

+++ pytorch_lamb/lamb16.py
@@ -85,9 +85,9 @@ class Lamb16(Optimizer):
                 if len(state) == 0:
                     state['step'] = 0
                     # Exponential moving average of gradient values
-                    state['exp_avg'] = torch.zeros_like(p.data, dtype=torch.float16)
+                    state['exp_avg'] = torch.zeros_like(p.data, dtype=torch.float8_e4m3fn)
                     # Exponential moving average of squared gradient values
-                    state['exp_avg_sq'] = torch.zeros_like(p.data, dtype=torch.float16)
+                    state['exp_avg_sq'] = torch.zeros_like(p.data, dtype=torch.float8_e4m3fn)
                     state['exp_avg_norm'] = 1.0
                     state['exp_avg_sq_norm'] = 1.0
